refactor(engine): TourProcessor + tour-estimate unification pin (PR-7)
Phase 2 engine-seams PR-7.

Part 1 — extract TourProcessor (no behavior change):
Verbatim move of three GameEngine methods into
shared/engine/processors/TourProcessor.ts:
  - processUnifiedTourRevenue (includes the PR-1 seeded varianceFactor)
  - getVenueNameFromAccess
  - applyTourPerformanceImpacts
Bodies are character-for-character identical, only this.* rebound to the
injected WeekContext (gameData/gameState/financialSystem/storage/getRandom)
and the two intra-class calls rebound. The engine keeps a same-signature
processUnifiedTourRevenue delegate so advanceProjectStages (stays in the
engine until PR-9) is unchanged. game-engine.ts: +9 / -261.

Part 2 — tour-estimate unification (investigation verdict: already true):
POST /api/tour/estimate and the engine tour path both assemble a
TourCalculationParams object and call the SINGLE FinancialSystem entry point
calculateDetailedTourBreakdown, which owns all tour math. No duplicated
assembly to extract; the differences are context-driven input sourcing only.
Added a clarifying comment in server/routes/tour.ts pointing at the shared
entry point + TourProcessor. No route behavior change.

Characterization: added
tests/endpoints/tour-estimate.characterization.test.ts pinning the full
happy-path estimate response. The calculateDetailedTourBreakdown core is
pinned exactly (explicit venueCapacity avoids the tier-fallback RNG draw);
the RNG-derived totalBudget/canAfford display fields are normalized out and
pinned structurally.

Gates: npm run check clean; golden master 8/8 twice with zero snapshot
updates (tour-week identical); tour-estimate + sweep-migrations pins green
twice; full suite 653/653.
